+++ video-call/src/main/java/com/nabla/sdk/videocall/NablaVideoCallClient.kt
@@ -1,0 +1,7 @@
+package com.nabla.sdk.videocall
+
+import android.content.Context
+
+public interface NablaVideoCallClient {
+    public fun openVideoCall(context: Context, url: String, roomId: String, token: String)
+}